feat: add cache for cv. Refactor: set the route to clear the entire cache
