+++ src/controllers/interest.controller.ts
@@ -112,19 +112,3 @@ export const deleteInterest = async (req: Request, res: Response): Promise<ApiRe
     return new ApiResponse(res, Errors.INTERNAL_SERVER_RESPONSE, e.message);
   }
 };
-
-export const cacheClear = async (_req: Request, res: Response): Promise<ApiResponse> => {
-  try {
-    if (cache.has('interests')) {
-      cache.del('interests');
-    }
-
-    return new ApiResponse(res, {
-      name: 'Success',
-      httpStatusCode: HttpStatusCodes.SUCCESS,
-      description: HttpStatusCodesDescriptions.SUCCESS,
-    });
-  } catch (e: any) {
-    return new ApiResponse(res, Errors.INTERNAL_SERVER_RESPONSE, e.message);
-  }
-};
