+++ src/routes/exhibitor.route.ts
@@ -6,7 +6,6 @@ import {
   addExhibitor,
   updateExhibitor,
   deleteExhibitor,
-  cacheClear,
 } from '../controllers/exhibitor.controller';
 import { isAuthenticated } from '../middlewares/auth.middleware';
 import { isInGroup } from '../middlewares/permission.middleware';
@@ -19,4 +18,3 @@ exhibitorRoutes.get('/exhibitor/:id', getExhibitor);
 exhibitorRoutes.post('/exhibitor', isAuthenticated, isInGroup([Roles.ADMIN]), addExhibitor);
 exhibitorRoutes.patch('/exhibitor/:id', isAuthenticated, isInGroup([Roles.ADMIN]), updateExhibitor);
 exhibitorRoutes.delete('/exhibitor/:id', isAuthenticated, isInGroup([Roles.ADMIN]), deleteExhibitor);
-exhibitorRoutes.get('/clear-exhibitor', isAuthenticated, isInGroup([Roles.ADMIN]), cacheClear);
